feat: 1차 추천 연산 기반 적재 마지막 스텝 완료

+++ src/main/java/com/example/masterplanbbe/infrastructure/repository/BatchRecommendationJdbcRepository.java
@@ -48,7 +48,7 @@ public class BatchRecommendationJdbcRepository {
                 rs.getLong("spec_id"),
                 rs.getLong("exam_id"),
                 rs.getString("age_group"),
-                rs.getInt("count_sum")), pageSize, offset);
+                rs.getDouble("count_sum")), pageSize, offset);
     }
 
     @Transactional
